tests: use kref_read() instead of accessing kref-internals
The v4.11 kernel introduced kref_read() for debugging and tests to read
refcount-values. Make use of it, rather than accessing the internals of
the refcounter.

Unfortunately, that means we need a hack to get kref_read() on older
kernels for now. We will remove it once 4.11 is released.

Note that we really *want* this debugging feature, since it is used
heavily to test the handle management. We rely on the ref-counter to
verify that handle acquisition works as expected.

This was reported by Tomáš Pružina.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>