peer: fix lifetime checks on RECV to use the anchor
When we dequeue messages during RECV, we discard anything that the
local peer has no access to, anymore. That is, if a peer explicitly
releases their last reference to a node, we will no longer return queued
messages for that node. This allows local nodes to re-use IDs and bind
their lifetime to memory addresses, if required.

However, the lifetime check in bus1_peer_peek() checked the sender
handle, rather than the receiver's node. Fix this to correctly
dereference the anchor of a destination handle.

This is similar to a recent fix, where we returned the handle-ID, rather
than the node-ID, which was fixed in:

    commit cf73d649bacac3ed81878f3e0d2c3133f0ad5929
    Author: David Herrmann <dh.herrmann@gmail.com>
    Date:   Sun Jan 15 10:40:55 2017 +0100

        peer: fix id-namespace used to return destination IDs on recv

Reported-by: Karl Nilsson <karl.nilsson.644@student.lu.se>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>